Modificação do README e Correção de Bug no momento do delete de um banner cujo imagem não estar presente no sistema

+++ models/Banners.php
@@ -92,6 +92,10 @@ class Banners extends ActiveRecord
 
     public function deleteBanner($file)
     {
+        if (!file_exists(\Yii::getAlias('@webroot/uploads/img/banners/').$file)) {
+            return false;
+        }
+        
         if (unlink(\Yii::getAlias('@webroot/uploads/img/banners/').$file)) {
             return true;
         } else {
